+++ CHANGE_LOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Create a plug-in which adds standard behavior for sources and javadoc jars
 - Create a plug-in which applies a configuration for specifying published information about a project
 - Create a plug-in which adds standard behavior for sources/JavaDoc artifacts in MavenPublications
+- Create a plug-in which applies publishedInfo to generated POM files